Move type safe index into common
This moves the type safe index template class into common, leaving the
multibody instantiations in place.
